fix #38
